requirements.txt
langroid>=0.1.12
chainlit>=0.7.0
weaviate-client>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
weaviate-client>=4.0.0